Return TEMPFAIL upon SERVFAIL
Currently, the milter is not implemented in a downgrade-resistant manner. In particular, an attacker can induce SERVFAIL responses, e.g. by performing a DoS attack on the authoritative server for the _smimecert subdomain. This will cause outbound mail to remain unencrypted. It is therefore better to treat a SERVFAIL response like failed DNSSEC authentication.